Add ORM\Query\QueryFactory.
