+++ tests/TestCase/ORM/QueryTest.php
@@ -3829,75 +3829,6 @@ class QueryTest extends TestCase
     }
 
     /**
-     * Tests subquery() copies connection by default.
-     */
-    public function testSubqueryConnection(): void
-    {
-        $subquery = Query::subquery($this->table);
-        $this->assertEquals($this->table->getConnection(), $subquery->getConnection());
-    }
-
-    /**
-     * Tests subquery() disables aliasing.
-     */
-    public function testSubqueryAliasing(): void
-    {
-        $articles = $this->getTableLocator()->get('Articles');
-        $subquery = Query::subquery($articles);
-
-        $subquery->select('Articles.field1');
-        $this->assertRegExpSql(
-            'SELECT <Articles>.<field1> FROM <articles> <Articles>',
-            $subquery->sql(),
-            !$this->connection->getDriver()->isAutoQuotingEnabled()
-        );
-
-        $subquery->select($articles, true);
-        $this->assertEqualsSql('SELECT id, author_id, title, body, published FROM articles Articles', $subquery->sql());
-
-        $subquery->selectAllExcept($articles, ['author_id'], true);
-        $this->assertEqualsSql('SELECT id, title, body, published FROM articles Articles', $subquery->sql());
-    }
-
-    /**
-     * Tests subquery() in where clause.
-     */
-    public function testSubqueryWhereClause(): void
-    {
-        $subquery = Query::subquery($this->getTableLocator()->get('Authors'))
-            ->select(['Authors.id'])
-            ->where(['Authors.name' => 'mariano']);
-
-        $query = $this->getTableLocator()->get('Articles')->find()
-            ->where(['Articles.author_id IN' => $subquery])
-            ->order(['Articles.id' => 'ASC']);
-
-        $results = $query->all()->toList();
-        $this->assertCount(2, $results);
-        $this->assertEquals([1, 3], array_column($results, 'id'));
-    }
-
-    /**
-     * Tests subquery() in join clause.
-     */
-    public function testSubqueryJoinClause(): void
-    {
-        $subquery = Query::subquery($this->getTableLocator()->get('Articles'))
-            ->select(['author_id']);
-
-        $query = $this->getTableLocator()->get('Authors')->find();
-        $query
-            ->select(['Authors.id', 'total_articles' => $query->func()->count('articles.author_id')])
-            ->leftJoin(['articles' => $subquery], ['articles.author_id' => new IdentifierExpression('Authors.id')])
-            ->group(['Authors.id'])
-            ->order(['Authors.id' => 'ASC']);
-
-        $results = $query->all()->toList();
-        $this->assertEquals(1, $results[0]->id);
-        $this->assertEquals(2, $results[0]->total_articles);
-    }
-
-    /**
      * Tests that queries that fetch associated data in separate queries do properly
      * inherit the hydration and results casting mode of the parent query.
      */
